Rename lighty-rnc packages
